dia api working
